requirements.txt
Django==2.0.6
django-rest==0.0.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
PyJWT==1.6.4
pytz==2018.4
